Registration of Agents rectified

+++ src/TranslationProvideInfo.java
@@ -6,8 +6,8 @@ public class TranslationProvideInfo implements Serializable, Cloneable {
 	private static final long serialVersionUID = 8246364560313449848L;
 	public TranslationService.Language language;
 	private final AID translatorAgent;
-	private float price;
-	public TranslationProvideInfo(TranslationService.Language lan, AID theAgent, float payment)
+	private double price;
+	public TranslationProvideInfo(TranslationService.Language lan, AID theAgent, double payment)
 	{	
 		super();
 		this.language = lan;
@@ -20,7 +20,7 @@ public class TranslationProvideInfo implements Serializable, Cloneable {
 	 public TranslationService.Language getLanguage() {
 		    return language;
 	}
-	 public float getPrice()
+	 public double getPrice()
 	 {
 		 return price;
 	 }
